Fixed agent crash when kernel supports unknown capabilities.
When capabilities are enabled, the capabilities initialisation code
would check that /proc/sys/kernel/cap_last_cap is less than
MAX_CAPABILITY, i.e. that the kernel doesn't support any capability the
code doesn't expect.  However the error message attempted to format
cap_last_cap value as a Capability enum, which would crash.

Part of fix for #10203.

+++ src/linux/capabilities.cpp
@@ -258,7 +258,7 @@ Try<Capabilities> Capabilities::create()
     return Error(
         "System last capability value '" + stringify(lastCap.get()) +
         "' is greater than maximum supported number of capabilities '" +
-        stringify(MAX_CAPABILITY) + "'");
+        stringify(static_cast<int>(MAX_CAPABILITY)) + "'");
   }
 
   // Test whether the kernel supports ambinent capabilities by testing
